# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Ember local-first productivity API — core loop for CLI and MCP clients.
servers:
  - url: /api
    description: Base API path (proxied from frontend; default http://localhost:8080/api)
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for a project
      parameters:
        - name: project_id
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update or action a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
...
  schemas:
...
    TaskStatus:
      type: string
      enum: [active, backlog, completed]
    TaskSource:
      type: string
      enum: [ai, user, reflection]
    Task:
      type: object
      properties:
        id: { type: string }
        project_id: { type: string }
        task_text: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        priority: { type: integer }
        task_order: { type: integer }
        source: { $ref: "#/components/schemas/TaskSource" }
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
      required:
        - id
        - project_id
        - task_text
        - status
        - priority
        - task_order
        - source
        - created_at
    CreateTaskRequest:
      type: object
      properties:
        project_id: { type: string }
        task_text: { type: string }
        source: { $ref: "#/components/schemas/TaskSource" }
        status: { $ref: "#/components/schemas/TaskStatus" }
      required: [project_id, task_text]
    UpdateTaskRequest:
      type: object
      properties:
        action:
          type: string
          enum: [complete, move_to_active, move_to_backlog]
        task_text: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
      required: [success]
...